+++ src/pages/LoginPage.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 import AuthLayout from '@/components/common/auth/AuthLayout';
-// import LoginForm from '../components/login/LoginForm';
-// import LinkToSignupPage from '../components/login/LinkToSignupPage';
+import LoginForm from '../components/login/LoginForm';
+import LinkToSignupPage from '../components/login/LinkToSignupPage';
 
 const LoginPage = () => (
   <AuthLayout>
-    <div>1</div>
-    {/* <LoginForm />
-    <LinkToSignupPage /> */}
+    <LoginForm />
+    <LinkToSignupPage />
   </AuthLayout>
 );
